models/Transformer/requirements.txt
python-socketio>=5.7.2
python-dotenv>=0.21.0
requests>=2.28.1
python-socketio[client]>=5.7.2
torch>=1.13.0